Progress on M x N, variable strike length tic-tac-toe player.
The plan is to create, using only vanilla JS, a tic-tac-toe player
that can learn to play several variants of the 2D tic-tac-toe game.
The variants are created by:

   (1)  varying the dimensions of the board to be M x N, where
        M >= 3 and N >=3;
   (2)  varying the strike length L, where 3 <= L <= M and
        3 <= L <= N;
   (3)  varying the strike overlap rule: whether you are
        allowed to create new stikes that contain a previously
        striked entry on the tic-tac-toe board;
   (4)  vary the threshold number of strikes that can be made
        by each player, such that above this threshold, the game
        shall be considered as "over".

Classic 2D tic-tac-toe is a special case of this: (M, N) = (3,3),
L = 3, and has a threshold of 1 strike. The overlapping strike
rule has no effect on a 3x3 2D tic-tac-toe.

The agents in this game use tabular Q-learning to learn 2D
variants of tic-tac-toe.

This is a squashed commit of 5 commits.